fix: primary key type

+++ src/main/java/com/danilo/ecommerce/repository/UserRepository.java
@@ -3,9 +3,9 @@ package com.danilo.ecommerce.repository;
 import com.danilo.ecommerce.domain.user.User;
 import org.springframework.data.repository.CrudRepository;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Optional;
 
-public interface UserRepository extends CrudRepository<User, BigDecimal> {
+public interface UserRepository extends CrudRepository<User, BigInteger> {
     Optional<User> findByUsername(String username);
 }
